memindah file dan menambahkan fitur sessionId

+++ server/api/auth/logout.post.js
@@ -5,11 +5,22 @@ export default defineEventHandler(async (event) => {
     const refreshToken = getCookie(event, "refreshToken");
 
     if (refreshToken) {
-      await prisma.refreshToken.deleteMany({
-        where: {
-          token: refreshToken,
-        },
+      const tokenInDb = await prisma.refreshToken.findUnique({
+        where: { token: refreshToken },
       });
+
+      if (tokenInDb) {
+        // Hapus refresh token
+        await prisma.refreshToken.deleteMany({
+          where: { token: refreshToken },
+        });
+
+        // Clear sessionId
+        await prisma.user.update({
+          where: { id: tokenInDb.userId },
+          data: { sessionId: null },
+        });
+      }
     }
 
     deleteCookie(event, "refreshToken");
@@ -20,6 +31,9 @@ export default defineEventHandler(async (event) => {
     console.log(error);
     deleteCookie(event, "refreshToken");
     setResponseStatus(event, 500);
-    return sendError(event, createError({statusCode: 500, message: "terjadi kesalahan server."}))
+    return sendError(
+      event,
+      createError({ statusCode: 500, message: "terjadi kesalahan server." }),
+    );
   }
 });
